requirements.txt

llama-index-embeddings-langchain
langchain
langchain-openai
nest_asyncio
pyvis
scikit-learn
python-dotenv
typing-extensions
scipy

llama-index==0.11.23
llama-index-core<0.12.0,>=0.11.23
llama-index-agent-openai==0.3.4

llama-index-vector-stores-pinecone>=0.2.0
gradio==4.44.1
mediapipe==0.10.15
pinecone==5.4.2
numpy<2.0,>=1.22.4
pillow<11.0
langchain==0.3.7

matplotlib<3.9.0,>=3.8.0
pandas>=1.5.0,<2.0

